+++ cloud/acp-spring-cloud-starter/src/main/kotlin/pers/acp/spring/cloud/AcpCloudComponentAutoConfiguration.kt
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.config.ConfigurableBeanFactory
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean
 import org.springframework.cloud.commons.util.InetUtils
 import org.springframework.context.annotation.*
+import pers.acp.spring.boot.component.ServerTools
 import pers.acp.spring.cloud.aspect.RestControllerRepeatAspect
 import pers.acp.spring.cloud.component.CloudTools
 import pers.acp.spring.cloud.error.AuthAccessDeniedHandler
@@ -34,7 +35,7 @@ class AcpCloudComponentAutoConfiguration {
     fun authExceptionEntryPoint(objectMapper: ObjectMapper) = AuthExceptionEntryPoint(objectMapper)
 
     @Bean
-    fun cloudIpTools(inetUtils: InetUtils) = CloudTools(inetUtils)
+    fun cloudIpTools(serverTools: ServerTools, inetUtils: InetUtils) = CloudTools(serverTools, inetUtils)
 
     @Bean
     @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)